Moved more functions from misc1.c to indent.c
