+++ src/misc1.h
@@ -25,7 +25,6 @@ void del_lines(long nlines, int undo);
 int gchar_pos(pos_T *pos);
 int gchar_cursor(void);
 void pchar_cursor(int c);
-int inindent(int extra);
 char_u *skip_to_option_part(char_u *p);
 void changed(void);
 void changed_int(void);
@@ -80,8 +79,6 @@ char_u *concat_fnames(char_u *fname1, char_u *fname2, int sep);
 char_u *concat_str(char_u *str1, char_u *str2);
 void add_pathsep(char_u *p);
 char_u *FullName_save(char_u *fname, int force);
-int get_expr_indent(void);
-int get_lisp_indent(void);
 void prepare_to_exit(void);
 void preserve_exit(void);
 int vim_fexists(char_u *fname);
