+++ src/search.c
@@ -21,6 +21,7 @@
 #include "fileio.h"
 #include "fold.h"
 #include "getchar.h"
+#include "indent.h"
 #include "main.h"
 #include "mark.h"
 #include "mbyte.h"
